requirements.txt
# Core Dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0
langgraph>=0.0.40
groq>=0.4.0

# Evaluation Metrics
rouge-score>=0.1.2
nltk>=3.8.1
bert-score>=0.3.13
sacrebleu>=2.4.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0

# API & Utilities
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0
tiktoken>=0.5.0

# Web Search (Serper)
google-search-results>=2.4.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Statistics
scipy>=1.11.0
scikit-learn>=1.3.0
